Update: move mol_dyn_result to handle_types so that it can use structure
