+++ enzy_htp/_interface/handle_types/__init__.py
@@ -9,3 +9,4 @@ Date: 2023-09-19
 
 from .mol_dyn_step import MolDynStep
 from .mol_dyn_parameterizer import MolDynParameterizer, MolDynParameter
+from .mol_dyn_result import MolDynResult, MolDynResultEgg
